backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.25.2
langchain
langchain-community
langchain-upstage
openai>=1.6.1
qdrant-client==1.7.0
neo4j==5.14.0
python-dotenv==1.0.0
langgraph==0.2.3
duckduckgo-search>=5.0.0
